dynamic model serializer
